# storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-garage Api",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://backend.egarage.store",
            "description": "Dev server online"
        },
        {
            "url": "http://localhost",
            "description": "Dev server offline"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/v1/last-parked-location/set": {
            "post": {
                "tags": [
                    "Last parked location management"
                ],
                "summary": "Set coordinates.",
                "parameters": [
                    {
                        "parameter": "user_credentials_in_query_required",
                        "name": "body",
                        "in": "query",
                        "description": "Coordinates of last parked location.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Coordinates"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/last-parked-location/": {
            "get": {
                "tags": [
                    "Last parked location management"
                ],
                "summary": "Get coordinates.",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Coordinates"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/last-parked-location/delete": {
            "delete": {
                "tags": [
                    "Last parked location management"
                ],
                "summary": "Delete coordinates.",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/account/update/password": {
            "put": {
                "tags": [
                    "Account Management"
                ],
                "summary": "Update user's account password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/account/update/name": {
            "put": {
                "tags": [
                    "Account Management"
                ],
                "summary": "Update user's account name",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateName"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/account/update/email": {
            "put": {
                "tags": [
                    "Account Management"
                ],
                "summary": "Update user's account email",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEmail"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/account/avatar/upload": {
            "post": {
                "tags": [
                    "Avatar Management"
                ],
                "summary": "Upload user's avatar",
                "parameters": [
                    {
                        "parameter": "user_credentials_in_query_required",
                        "name": "body",
                        "in": "query",
                        "description": "Avatar needed to perform action. Acceptable extensions: png, jpg, jpeg.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Avatar"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/account/avatar": {
            "get": {
                "tags": [
                    "Avatar Management"
                ],
                "summary": "Get user's avatar",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "image": {
                                "schema": {
                                    "$ref": "#/components/schemas/RetrievedAvatar"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/account/avatar/delete": {
            "delete": {
                "tags": [
                    "Avatar Management"
                ],
                "summary": "Delete user's avatar",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Operates about user",
                "responses": {
                    "201": {
                        "description": "Logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "message": "loged out"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Coordinates": {
                "properties": {
                    "longitude": {
                        "type": "numeric|required"
                    },
                    "latitude": {
                        "type": "numeric|required"
                    }
                },
                "type": "object",
                "example": {
                    "longitude": "+50.235423",
                    "latitude": "-15.152354"
                }
            },
            "UpdatePassword": {
                "properties": {
                    "password": {
                        "type": "string"
                    },
                    "password_confirmation": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "password": "12345678",
                    "password_confirmation": "12345678"
                }
            },
            "UpdateEmail": {
                "properties": {
                    "email": {
                        "type": "email|string"
                    }
                },
                "type": "object",
                "example": {
                    "email": "test@test.com"
                }
            },
            "UpdateName": {
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "name": "JohnDoe"
                }
            },
            "Avatar": {
                "properties": {
                    "image": {
                        "type": "file"
                    }
                },
                "type": "object",
                "example": {
                    "image": "file"
                }
            },
            "RetrievedAvatar": {
                "properties": {
                    "image": {
                        "description": "Image is encoded in base64.",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "image": "iVBORw0KGgoAAAANSUhEUgAAAFEAAABRCAYAAACqj0o2AAAACXBIWXMAAA7EAAAOxAGVKw4bAAAAzElEQVR4nO3QgQkAIRDAsNP9d/4fwoIgyQSla2a+4ci+HfACEwMmBkwMmBgwMWBiwMSAiQETAyYGTAyYGDAxYGLAxICJARMDJgZMDJgYMDFgYsDEgIkBEwMmBkwMmBgwMWBiwMSAiQETAyYGTAyYGDAxYGLAxICJARMDJgZMDJgYMDFgYsDEgIkBEwMmBkwMmBgwMWBiwMSAiQETAyYGTAyYGDAxYGLAxICJARMDJgZMDJgYMDFgYsDEgIkBEwMmBkwMmBgwMWBiwMTAD2JSAaEhoH6NAAAAAElFTkSuQmCC"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}